chore: remove debug logs refactor: separate webworker functions to separate modules
