+++ src/utils/plots/pca.ts
@@ -1,19 +1,24 @@
-import { dataTable } from '@/store/data-store';
+import { DataRow } from '@/store/dataframe';
 import { PCA } from 'ml-pca';
 
 import { Layout, PlotData } from 'plotly.js';
+import { getSubheaderColors } from '../color';
+import { toArrayOfColumns } from '../store';
 
 const sprintfNum = (num: number): string => {
   return (Math.round(num * 1000) / 1000).toFixed(3);
 };
 
-const pcaData = async (
-  title?: string
-): Promise<{
+const pcaData = (
+  rows: DataRow,
+  srcReplicateNames: string[],
+  multiHeaderSep: string,
+  plotTitle?: string
+): {
   data: Partial<PlotData>[];
   layout: Partial<Layout>;
-}> => {
-  const data2dArr = dataTable.toArrayOfColumns();
+} => {
+  const data2dArr = toArrayOfColumns(rows);
   const pca = new PCA(data2dArr);
   // Project the data2dArr into PC coordinate system:
   const projectedData = pca.predict(data2dArr);
@@ -24,14 +29,14 @@ const pcaData = async (
       y: projectedData.getColumn(1),
       type: 'scatter',
       mode: 'markers',
-      text: dataTable.colNames,
+      text: srcReplicateNames,
       textfont: {
         family: 'Times New Roman',
       },
       textposition: 'bottom center',
       marker: {
         size: 12,
-        color: dataTable.getSubheaderColors(2),
+        color: getSubheaderColors(2, srcReplicateNames, multiHeaderSep),
       },
     },
   ];
@@ -39,7 +44,7 @@ const pcaData = async (
   const varExpl = pca.getExplainedVariance();
   const layout: Partial<Layout> = {
     title: {
-      text: title,
+      text: plotTitle,
       font: {
         size: 20,
       },
